Price now calculated. Result appears to be incorrect however.

+++ ADPROC_Coursework/src/adproc_coursework/LongPipe.java
@@ -30,7 +30,7 @@ public abstract class LongPipe {
 		grade = newGrade;
 	}
 
-	public double calculateVolume(int newLength, int newDiameter) {
+	public double calculateVolume(float newLength, float newDiameter) {
 		//The volume of a cylinder = V=πr^2h
 		double totalVolume;
 		double innerVolume;
